requirements.txt
-i https://pypi.org/simple
amqp==5.3.1; python_version >= '3.6'
anyio==4.9.0; python_version >= '3.9'
asgiref==3.8.1; python_version >= '3.8'
attrs==25.3.0; python_version >= '3.8'
billiard==4.2.1; python_version >= '3.7'
celery==5.5.3; python_version >= '3.8'
certifi==2025.4.26; python_version >= '3.6'
click==8.2.1; python_version >= '3.10'
click-didyoumean==0.3.1; python_full_version >= '3.6.2'
click-plugins==1.1.1
click-repl==0.3.0; python_version >= '3.6'
colorama==0.4.6; platform_system == 'Windows'
django==5.2.2; python_version >= '3.10'
django-redis==5.4.0; python_version >= '3.6'
djangorestframework==3.16.0; python_version >= '3.9'
drf-spectacular==0.28.0; python_version >= '3.7'
h11==0.16.0; python_version >= '3.8'
httpcore==1.0.9; python_version >= '3.8'
httpx==0.28.1; python_version >= '3.8'
idna==3.10; python_version >= '3.6'
inflection==0.5.1; python_version >= '3.5'
jsonschema==4.24.0; python_version >= '3.9'
jsonschema-specifications==2025.4.1; python_version >= '3.9'
kombu==5.5.4; python_version >= '3.8'
packaging==25.0; python_version >= '3.8'
prompt-toolkit==3.0.51; python_version >= '3.8'
psycopg[binary]==3.2.9; python_version >= '3.8'
psycopg-binary==3.2.9
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-decouple==3.8
pyyaml==6.0.2; python_version >= '3.8'
redis==6.2.0; python_version >= '3.9'
referencing==0.36.2; python_version >= '3.9'
rpds-py==0.25.1; python_version >= '3.9'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.1; python_version >= '3.7'
sqlparse==0.5.3; python_version >= '3.8'
typing-extensions==4.14.0; python_version < '3.13'
tzdata==2025.2; sys_platform == 'win32'
uritemplate==4.2.0; python_version >= '3.9'
vine==5.1.0; python_version >= '3.6'
wcwidth==0.2.13
